EO-894 - Rename new Alerts pages
