+++ src/pages/alerts/EditPage.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import { Page } from '@sparkpost/matchbox';
-import withEditPage from './containers/EditPageNew.container';
+import withEditPage from './containers/EditPage.container';
 import AlertFormNew from './components/AlertFormNew';
 import formatFormValues from './helpers/formatFormValues';
 import { Loading } from 'src/components';
 import RedirectAndAlert from 'src/components/globalAlert/RedirectAndAlert';
 
-export class EditPageNew extends Component {
+export class EditPage extends Component {
 
   componentDidMount() {
     const { getAlert, id } = this.props;
@@ -55,4 +55,4 @@ export class EditPageNew extends Component {
   }
 }
 
-export default withEditPage(EditPageNew);
+export default withEditPage(EditPage);
